+++ src/main/scala/leapfrogTriejoin/implicits.scala
@@ -1,12 +1,12 @@
 package leapfrogTriejoin
 
 object implicits {
-  implicit def leapfrogTriejoin2Iterator(lftj: LeapfrogTriejoin): Iterator[Array[Int]] = {
-    new Iterator[Array[Int]] {
+  implicit def leapfrogTriejoin2Iterator(lftj: LeapfrogTriejoin): Iterator[Array[Long]] = {
+    new Iterator[Array[Long]] {
       override def hasNext: Boolean = !lftj.atEnd
 
 
-      override def next(): Array[Int] = {
+      override def next(): Array[Long] = {
         lftj.next()
       }
     }
